refactor: survey 페이지 SurveyContent 컴포넌트 분리 #161
